번개 생성 시 이미지 값 처리, 번개 리스트 response 데이터 추가, 번개 삭제 시 오류 수정

+++ src/service/lightService.js
@@ -5,6 +5,7 @@ const responseMessage = require('../constants/responseMessage');
 const db = require('../loaders/db');
 const dayjs = require('dayjs');
 const { calculateAge } = require('../lib/calculateAge');
+const { applyKoreanTime } = require('../lib/applyKoreanTime');
 
 const addLight = async (category, title, date, place, people_cnt, description, image, organizerId, crewId, time) => {
   let client;
@@ -23,14 +24,13 @@ const addLight = async (category, title, date, place, people_cnt, description, i
 
     // 존재 하는 동아리인지 검사
     const existCrew = await crewDao.getExistCrew(client, crewId);
-    if(!existCrew) {
+    if (!existCrew) {
       return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_CREW);
     }
 
-    const data = await lightDao.addLight(client, category, title, date, place
-        , people_cnt, description,image, organizerId, crewId, time);
+    const data = await lightDao.addLight(client, category, title, date, place, people_cnt, description, image, organizerId, crewId, time);
     const result = [data];
-    const data_result = result.map(o => ({
+    const data_result = result.map((o) => ({
       id: Number(o.id),
       category: o.category,
       title: o.title,
@@ -40,12 +40,12 @@ const addLight = async (category, title, date, place, people_cnt, description, i
       description: o.description,
       image: o.image,
       isDeleted: o.isDeleted,
-      createdAt: o.createdAt,
-      updatedAt: o.updatedAt,
+      createdAt: applyKoreanTime(o.createdAt),
+      updatedAt: applyKoreanTime(o.updatedAt),
       organizerId: Number(o.organizerId),
       crewId: Number(o.crewId),
-      time: o.time
-  }));
+      time: o.time,
+    }));
 
     // 번개 생성 후 번개 소유자도 번개에 참여시키기
     await lightDao.addLightOrganizer(client, organizerId);
@@ -54,7 +54,7 @@ const addLight = async (category, title, date, place, people_cnt, description, i
     return util.success(statusCode.OK, responseMessage.LIGHT_ADD_SUCCESS, data_result);
   } catch (error) {
     await client.query('ROLLBACK');
-    console.log('addLight에서 error 발생'+ error);
+    console.log('addLight에서 error 발생' + error);
   } finally {
     client.release();
   }
@@ -69,8 +69,8 @@ const putLight = async (lightId, organizerId, category, title, date, place, peop
     client = await db.connect(log);
     await client.query('BEGIN');
     // 존재하는 번개인지 확인
-    const existLight = await lightDao.getExistLight(client,lightId);
-    if(!existLight) {
+    const existLight = await lightDao.getExistLight(client, lightId);
+    if (!existLight) {
       return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_LIGHT);
     }
     // 존재하는 유저인지 확인
@@ -79,39 +79,38 @@ const putLight = async (lightId, organizerId, category, title, date, place, peop
       return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_USER);
     }
     // 번개 수정하려는 사람이 organizer인지 검사
-    const organizer = await lightDao.getLightOrganizerById(client,organizerId);
+    const organizer = await lightDao.getLightOrganizerById(client, organizerId);
     if (!organizer) {
       return util.fail(statusCode.BAD_REQUEST, responseMessage.NOT_LIGHT_ORGANIZER);
     }
-    const data = await lightDao.putLight(client,lightId,organizerId,category,title, date, place
-      , people_cnt, description, time);
+    const data = await lightDao.putLight(client, lightId, organizerId, category, title, date, place, people_cnt, description, time);
     const result = [data];
-    const data_result = result.map(o => ({
-        id: Number(o.id),
-        category: o.category,
-        title: o.title,
-        date: o.date,
-        place: o.place,
-        peopleCnt: o.peopleCnt,
-        description: o.description,
-        image: o.image,
-        isDeleted: o.isDeleted,
-        createdAt: o.createdAt,
-        updatedAt: o.updatedAt,
-        organizerId: Number(o.organizerId),
-        crewId: Number(o.crewId),
-        time: o.time
+    const data_result = result.map((o) => ({
+      id: Number(o.id),
+      category: o.category,
+      title: o.title,
+      date: o.date,
+      place: o.place,
+      peopleCnt: o.peopleCnt,
+      description: o.description,
+      image: o.image,
+      isDeleted: o.isDeleted,
+      createdAt: applyKoreanTime(o.createdAt),
+      updatedAt: applyKoreanTime(o.updatedAt),
+      organizerId: Number(o.organizerId),
+      crewId: Number(o.crewId),
+      time: o.time,
     }));
-    
+
     await client.query('COMMIT');
     return util.success(statusCode.OK, responseMessage.LIGHT_PUT_SUCCESS, data_result);
   } catch (error) {
     await client.query('ROLLBACK');
-    console.log('putLight에서 error 발생'+ error);
+    console.log('putLight에서 error 발생' + error);
   } finally {
     client.release();
   }
-}
+};
 const postEnterLight = async (lightId, memberId) => {
   let client;
 
@@ -120,8 +119,8 @@ const postEnterLight = async (lightId, memberId) => {
     client = await db.connect(log);
     await client.query('BEGIN');
     // 존재하는 번개인지 확인
-    const existLight = await lightDao.getExistLight(client,lightId);
-    if(!existLight) {
+    const existLight = await lightDao.getExistLight(client, lightId);
+    if (!existLight) {
       return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_LIGHT);
     }
     // 존재하는 유저인지 확인
@@ -129,17 +128,17 @@ const postEnterLight = async (lightId, memberId) => {
     if (!exist) {
       return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_USER);
     }
-    await lightUserDao.postEnterLight(client,lightId, memberId);
-    
+    await lightUserDao.postEnterLight(client, lightId, memberId);
+
     await client.query('COMMIT');
     return util.success(statusCode.OK, responseMessage.LIGHT_ENTER_SUCCESS);
   } catch (error) {
     await client.query('ROLLBACK');
-    console.log('postEnterLight error 발생'+ error);
+    console.log('postEnterLight error 발생' + error);
   } finally {
     client.release();
   }
-}
+};
 const deleteCancelLight = async (lightId, memberId) => {
   let client;
 
@@ -148,21 +147,21 @@ const deleteCancelLight = async (lightId, memberId) => {
     client = await db.connect(log);
     await client.query('BEGIN');
     // 번개에 참여한 유저인지 검사
-    const enterLightMember = await lightUserDao.getEnterLightMember(client,lightId, memberId);
-    if(!enterLightMember){
+    const enterLightMember = await lightUserDao.getEnterLightMember(client, lightId, memberId);
+    if (!enterLightMember) {
       return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_LIGHT_MEMBER);
     }
-    await lightUserDao.deleteCancelLight(client,lightId, memberId);
-    
+    await lightUserDao.deleteCancelLight(client, lightId, memberId);
+
     await client.query('COMMIT');
     return util.success(statusCode.OK, responseMessage.LIGHT_CANCEL_SUCCESS);
   } catch (error) {
     await client.query('ROLLBACK');
-    console.log('deleteCancelLight error 발생'+ error);
+    console.log('deleteCancelLight error 발생' + error);
   } finally {
     client.release();
   }
-}
+};
 const getEnterLightMember = async (lightId, memberId) => {
   let client;
 
@@ -170,13 +169,13 @@ const getEnterLightMember = async (lightId, memberId) => {
   try {
     client = await db.connect(log);
     const data = await lightUserDao.getEnterLightMember(client, lightId, memberId);
-    return data
+    return data;
   } catch (error) {
-    console.log('getEnterLightMember error 발생'+ error);
+    console.log('getEnterLightMember error 발생' + error);
   } finally {
     client.release();
   }
-}
+};
 const deleteLight = async (lightId, organizerId) => {
   let client;
 
@@ -185,21 +184,28 @@ const deleteLight = async (lightId, organizerId) => {
     client = await db.connect(log);
     await client.query('BEGIN');
     // 번개 삭제하려는 사람이 organizer인지 검사
-    const organizer = await lightDao.getLightOrganizerById(client, organizerId);
+    const organizer = await lightDao.getLightOrganizerById(client, lightId,organizerId);
     if (!organizer) {
       return util.fail(statusCode.BAD_REQUEST, responseMessage.NOT_LIGHT_ORGANIZER);
     }
+    // 존재하는 번개인지 확인
+    const existLight = await lightDao.getExistLight(client, lightId);
+    if (!existLight) {
+      return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_LIGHT);
+    }
+    // 번개 삭제를 하려면 그 번개에 속해있는 사람 전부 제거.
+    await lightUserDao.deleteLightUser(client, lightId);
     await lightDao.deleteLight(client, lightId, organizerId);
-    
+
     await client.query('COMMIT');
     return util.success(statusCode.OK, responseMessage.LIGHT_DELETE_SUCCESS);
   } catch (error) {
     await client.query('ROLLBACK');
-    console.log('deleteLight error 발생'+ error);
+    console.log('deleteLight error 발생' + error);
   } finally {
     client.release();
   }
-}
+};
 const getOrganizerLight = async (organizerId) => {
   let client;
 
@@ -214,26 +220,26 @@ const getOrganizerLight = async (organizerId) => {
     }
 
     const result = await lightDao.getOrganizerLight(client, organizerId);
-    const data = result.map(light => ({
-        light_id : Number(light.id),
-        title: light.title,
-        category: light.category,
-        date: dayjs(light.date).format('YYYY-MM-DD'),
-        time: light.time.slice(0,-3),
-        people_cnt: light.peopleCnt,
-        place: light.place,
-        LightMemberCnt: Number(light.joinCnt)
-    }))
+    const data = result.map((light) => ({
+      light_id: Number(light.id),
+      title: light.title,
+      category: light.category,
+      date: dayjs(light.date).format('YYYY-MM-DD'),
+      time: light.time.slice(0, -3),
+      people_cnt: light.peopleCnt,
+      place: light.place,
+      LightMemberCnt: Number(light.joinCnt),
+    }));
 
     await client.query('COMMIT');
     return util.success(statusCode.OK, responseMessage.LIGHT_GET_ORGANIZER_SUCCESS, data);
   } catch (error) {
     await client.query('ROLLBACK');
-    console.log('getOrganizerLight error 발생'+ error);
+    console.log('getOrganizerLight error 발생' + error);
   } finally {
     client.release();
   }
-}
+};
 const getEnterLight = async (memberId) => {
   let client;
 
@@ -241,7 +247,7 @@ const getEnterLight = async (memberId) => {
   try {
     client = await db.connect(log);
     await client.query('BEGIN');
-    
+
     // 존재하는 유저인지 확인
     const exist = await userDao.getUserById(client, memberId);
     if (!exist) {
@@ -249,26 +255,26 @@ const getEnterLight = async (memberId) => {
     }
 
     const result = await lightDao.getEnterLight(client, memberId);
-    const data = result.map(light => ({
-        light_id : Number(light.id),
-        title: light.title,
-        category: light.category,
-        date: dayjs(light.date).format('YYYY-MM-DD'),
-        time: light.time.slice(0,-3),
-        people_cnt: light.peopleCnt,
-        place: light.place,
-        LightMemberCnt: Number(light.joinCnt)
-    }))
+    const data = result.map((light) => ({
+      light_id: Number(light.id),
+      title: light.title,
+      category: light.category,
+      date: dayjs(light.date).format('YYYY-MM-DD'),
+      time: light.time.slice(0, -3),
+      people_cnt: light.peopleCnt,
+      place: light.place,
+      LightMemberCnt: Number(light.joinCnt),
+    }));
 
     await client.query('COMMIT');
     return util.success(statusCode.OK, responseMessage.LIGHT_GET_ENTER_SUCCESS, data);
   } catch (error) {
     await client.query('ROLLBACK');
-    console.log('getEnterLight error 발생'+ error);
+    console.log('getEnterLight error 발생' + error);
   } finally {
     client.release();
   }
-}
+};
 const getScrapLight = async (memberId) => {
   let client;
 
@@ -284,55 +290,55 @@ const getScrapLight = async (memberId) => {
     }
 
     const result = await lightDao.getScrapLight(client, memberId);
-    const data = result.map(light => ({
-        light_id : Number(light.id),
-        title: light.title,
-        category: light.category,
-        date: dayjs(light.date).format('YYYY-MM-DD'),
-        time: light.time.slice(0,-3),
-        people_cnt: light.peopleCnt,
-        place: light.place,
-        LightMemberCnt: Number(light.joinCnt)
-    }))
+    const data = result.map((light) => ({
+      light_id: Number(light.id),
+      title: light.title,
+      category: light.category,
+      date: dayjs(light.date).format('YYYY-MM-DD'),
+      time: light.time.slice(0, -3),
+      people_cnt: light.peopleCnt,
+      place: light.place,
+      LightMemberCnt: Number(light.joinCnt),
+    }));
     await client.query('COMMIT');
     return util.success(statusCode.OK, responseMessage.LIGHT_GET_SCRAP_SUCCESS, data);
   } catch (error) {
     await client.query('ROLLBACK');
-    console.log('getScrapLight error 발생'+ error);
+    console.log('getScrapLight error 발생' + error);
   } finally {
     client.release();
   }
-}
+};
 const getCategoryLight = async (category, sort) => {
   let client;
 
   const log = `lightService.getCategoryLight | category = ${category}, sort = ${sort}`;
   // 카테고리가 없으면 없는 번개
-  if(!category){
+  if (!category) {
     return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_LIGHT);
   }
   try {
     client = await db.connect(log);
-    const result = await lightDao.getCategoryLight(client, category,sort);
+    const result = await lightDao.getCategoryLight(client, category, sort);
 
-    const data = result.map(light => ({
-      light_id : Number(light.id),
+    const data = result.map((light) => ({
+      light_id: Number(light.id),
       category: light.category,
       title: light.title,
       date: dayjs(light.date).format('YYYY-MM-DD'),
-      time: light.time.slice(0,-3),
+      time: light.time.slice(0, -3),
       people_cnt: light.peopleCnt,
       place: light.place,
-      LightMemberCnt: Number(light.joinCnt)
-    }))
+      LightMemberCnt: Number(light.joinCnt),
+    }));
 
     return util.success(statusCode.OK, responseMessage.LIGHT_GET_CATEGORY_SUCCESS, data);
   } catch (error) {
-    console.log('getCategoryLight error 발생'+ error);
+    console.log('getCategoryLight error 발생' + error);
   } finally {
     client.release();
   }
-}
+};
 const getLightDetail = async (lightId) => {
   let client;
 
@@ -342,62 +348,62 @@ const getLightDetail = async (lightId) => {
     await client.query('BEGIN');
     // 존재하는 번개인지 확인
     const existLight = await lightDao.getExistLight(client, lightId);
-    if(!existLight) {
+    if (!existLight) {
       return util.fail(statusCode.BAD_REQUEST, responseMessage.NO_LIGHT);
     }
     const result = await lightDao.getLightDetail(client, lightId);
     const members = await lightDao.getLightDetailMember(client, lightId);
     const organizer = await lightDao.getLightDetailOrganizer(client, lightId);
-    
-    const data2 = members.map(o => ({
+
+    const data2 = members.map((o) => ({
       user_id: Number(o.id),
-      mbti : o.mbti,
+      mbti: o.mbti,
       gender: o.gender,
-      name : o.name,
-      age : Number(calculateAge(dayjs(o.birthDay).format('YYYY-MM-DD'))) + 1
-  }))
+      name: o.name,
+      age: Number(calculateAge(dayjs(o.birthDay).format('YYYY-MM-DD'))) + 1,
+    }));
 
-    const data3 = organizer.map(o => ({
+    const data3 = organizer.map((o) => ({
+      userLoginId: o.userLoginId,
       organizer_id: Number(o.id),
-      name: o.name
-    })) 
+      name: o.name,
+    }));
 
-    const data = result.map(light => ({
-        light_id : Number(light.id),
-        category : light.category,
-        title: light.title,
-        date: dayjs(light.date).format('YYYY-MM-DD'),
-        time: light.time.slice(0,-3),
-        description: light.description,
-        image: light.image,
-        people_cnt: light.peopleCnt,
-        place: light.place,
-        LightMemberCnt: Number(light.joinCnt),
-        members: data2,
-        organizer: data3
-    }))
+    const data = result.map((light) => ({
+      light_id: Number(light.id),
+      category: light.category,
+      title: light.title,
+      date: dayjs(light.date).format('YYYY-MM-DD'),
+      time: light.time.slice(0, -3),
+      description: light.description,
+      image: light.image,
+      people_cnt: light.peopleCnt,
+      place: light.place,
+      LightMemberCnt: Number(light.joinCnt),
+      members: data2,
+      organizer: data3,
+    }));
 
     await client.query('COMMIT');
     return util.success(statusCode.OK, responseMessage.LIGHT_GET_DETAIL_SUCCESS, data);
   } catch (error) {
     await client.query('ROLLBACK');
-    console.log('getLightDetail error 발생'+ error);
+    console.log('getLightDetail error 발생' + error);
   } finally {
     client.release();
   }
-}
-
+};
 
 module.exports = {
-    addLight,
-    putLight,
-    postEnterLight,
-    deleteCancelLight,
-    getEnterLightMember,
-    deleteLight,
-    getOrganizerLight,
-    getEnterLight,
-    getScrapLight,
-    getCategoryLight,
-    getLightDetail
+  addLight,
+  putLight,
+  postEnterLight,
+  deleteCancelLight,
+  getEnterLightMember,
+  deleteLight,
+  getOrganizerLight,
+  getEnterLight,
+  getScrapLight,
+  getCategoryLight,
+  getLightDetail,
 };
